+++ Modules/antibodies/Model/AcStaining.php
@@ -94,6 +94,7 @@ class AcStaining extends Model {
         $sql = "insert into ac_stainings(name, id_space)"
                 . " values(?,?)";
         $this->runRequest($sql, array($name, $id_space));
+        return $this->getDatabase()->lastInsertId();
     }
 
     /**