+++ Modules/antibodies/Model/Aciinc.php
@@ -84,6 +84,7 @@ class Aciinc extends Model {
         $sql = "insert into ac_aciincs(nom, id_space)"
                 . " values(?,?)";
         $this->runRequest($sql, array($name, $id_space));
+        return $this->getDatabase()->lastInsertId();
     }
 
     /**