+++ Modules/antibodies/Model/Dem.php
@@ -85,6 +85,7 @@ class Dem extends Model {
         $sql = "insert into ac_dems(nom, id_space)"
                 . " values(?,?)";
         $this->runRequest($sql, array($name, $id_space));
+        return $this->getDatabase()->lastInsertId();
     }
 
     /**